% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_fpts.R
\name{get_fpts_qb_season}
\alias{get_fpts_qb_season}
\title{QB Season Fantasy Points and Statistics}
\usage{
get_fpts_qb_season(pbp_db, szn = 2024)
}
\arguments{
\item{pbp_db}{Play-by-play database table}

\item{szn}{Season}
}
\value{
Obtain QB cumulative statistics and fantasy points for season
}
\description{
QB Season Fantasy Points and Statistics
}
