% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbp_data.R
\name{get_pbp_data}
\alias{get_pbp_data}
\title{Obtain Play-by-Play Data for Specified Time Frame}
\usage{
get_pbp_data(
  pbp_db = NULL,
  pbp_db_tbl = NULL,
  season = NULL,
  week_min = NULL,
  week_max = NULL
)
}
\arguments{
\item{pbp_db_tbl}{Play-by-Play database table name (optional)}

\item{season}{NFL season (required) to obtain play-by-play data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}

\item{week_min}{Minimum week (required) to define whether pulling a range
of weeks or the entire season. Use \code{week_min = 1} for the first week of
the season, must be an integer.}

\item{week_max}{Maximum week (optional) to define a range of weeks to pull
from an NFL season. If not defined, the data will be pulled for all weeks,
beginning with \code{week_min}.}

\item{pbp_dp}{Play-by-Play database path (optional)}
}
\value{
Dataframe with NFL play-by-play data
}
\description{
Obtain play-by-play data for a specified time frame from either a saved
database or if not defined, using \code{nflreadr::load_pbp()}.
}
\details{
The function \code{get_pbp_data} can be utilized to obtain play-by-play data
by either loading from a saved database or using \code{nflreadr::load_pbp()}.
The data is obtained for a user-defined season or multiple seasons.
A range of weeks can also be defined with \code{week_min} and \code{week_max}.
If the entire season is desired, use \code{week_min = 1} and \code{week_max} does
not need to be defined.
To specify loading from a database, define the path to the database with
\code{pbp_db} as well as the name of the table to load with \code{pbp_db_tbl}.
To load from a database, you will need to save play-by-play data to a
database using the \code{nflfastR} function, \code{update_db()}.
For example, the database is saved by default as \code{pbp_db} with a table that
is stored containing all play-by-play information called \code{nflfastR_pbp}.
Assume that the database is saved in \verb{project_name/data/}.
Using the default naming scheme, \code{get_pbp_data} can be defined using the
database with \code{pbp_db = "data/pbp_db"} and \code{pbp_db_tbl = "nflfastR_pbp"}.
Note that these two arguments must be defined as strings.
For more information, \code{nflfastR} provides an example on using the database
in \link{Example 8: Using the built-in database function}
(https://www.nflfastr.com/articles/nflfastR.html
#example-8-using-the-built-in-database-function)
}
\seealso{
\code{\link[nflreadr]{load_pbp}}
Load play-by-play data,

\code{\link[nflfastR]{update_db}}
Update or Create a nflfastR Play-by-Play Database
}
\author{
Nolan MacDonald

\itemize{
\item{\code{connect}}{
Connect to the play-by-play database with the \code{RSQLite} package
}
\item{\code{pbp_db}}{Load the play-by-play data from the database table}
\item{\code{pbp}}{Play-by-play dataframe filtered for season(s) and week(s)}
}
}
