% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_data.R
\name{get_player_data}
\alias{get_player_data}
\title{Player Identification Data}
\usage{
get_player_data(season = NULL, pos = NULL)
}
\arguments{
\item{season}{NFL season or year (int)}

\item{pos}{Player position (str) such as "WR"}
}
\value{
Dataframe that lists player data for a specified season and position
}
\description{
Obtain player identification data for various platforms.
}
\details{
Use the function \code{get_player_data} to obtain player information for a
defined season and position. Player information includes several IDs for
platforms such as ESPN, Rotowire, ESPN, etc.
}
\seealso{
\code{\link[nflreadr]{load_rosters}}
Load roster information,

\code{\link[nuclearff]{replace_player_names}}
Cleanup player names for continuity

\code{\link[nuclearff]{replace_team_names}}
Cleanup team abbreviations for continuity

@author Nolan MacDonald
}
