% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_stats_season.R
\name{get_qb_combined_stats_season}
\alias{get_qb_combined_stats_season}
\title{QB Cumulative Season Advanced Stats}
\format{
A data frame with 79 variables that are described below.
\describe{
\item{\code{player_id}}{Player gsis id (e.g., 00-0034796)}
\item{\code{player_display_name}}{Player name (e.g., Lamar Jackson)}
\item{\code{player_name}}{Player shortened name (e.g., L.Jackson)}
\item{\code{pos}}{Player position (e.g., QB)}
\item{\code{tm}}{Player team (e.g., BAL)}
\item{\code{g}}{Number of games played}
\item{\code{completions}}{Total pass completions (CMP)from PBP data}
\item{\code{attempts}}{Total pass attempts (ATT)from PBP data}
\item{\code{cmp_pct}}{Pass completion percentage from PBP data}
\item{\code{expected_completion_percentage}}{
Expected completion percentage from NGS
}
\item{\code{completion_percentage_above_expectation}}{
Average expected completion percentage (xCOMP) from NGS
}
\item{\code{passing_yards}}{Total passing yards from PBP data}
\item{\code{passing_tds}}{Total passing touchdowns from PBP data}
\item{\code{interceptions}}{Total pass interceptions (INT) from PBP data}
\item{\code{passing_epa}}{
Total expected points added (EPA) on pass attempts and sacks.
NOTE: This uses the variable \code{qb_epa}, which gives QB credit for EPA for
up to the point where a receiver lost a fumble after a completed catch
and makes EPA work more like passing yards on plays with fumbles
}
\item{\code{passer_rating}}{Average passer rating from NGS}
\item{\code{pacr}}{
Passing Air Conversion Ratio.
PACR = \code{passing_yards} / \code{passing_air_yards}  from PBP data
}
\item{\code{dakota}}{
Adjusted EPA + CPOE composite based on coefficients which best predicts
adjusted EPA/play in the following year from PBP data
}
\item{\code{aggressiveness}}{Average aggressiveness (AGG\%) from NGS}
\item{\code{passing_air_yards}}{
Passing air yards (includes incomplete passes) from PBP data
}
\item{\code{passing_yards_after_catch}}{
Yards after the catch gained on plays in which player was the passer
(this is an unofficial stat and may differ slightly between different
sources) from PBP data
}
\item{\code{avg_air_distance}}{Average air distance from NGS}
\item{\code{max_air_distance}}{Maximum or longest air distance from NGS}
\item{\code{avg_time_to_throw}}{Average time to throw (TT) from NGS}
\item{\code{avg_completed_air_yards}}{
Average completed air yards (CAY) from NGS
}
\item{\code{avg_intended_air_yards}}{
Average intended air yards (IAY) from NGS
}
\item{\code{avg_air_yards_differential}}{
Average air yards differential (AYD) from NGS
}
\item{\code{max_completed_air_distance}}{
Maximum or longest completed air distance (LCAD) from NGS
}
\item{\code{avg_air_yards_to_sticks}}{
Average air yards to the sticks (AYTS) from NGS
}
\item{\code{passing_first_downs}}{
First downs on pass attempts from PBP data
}
\item{\code{throwaways}}{Number of throwaways from PFR}
\item{\code{spikes}}{Number of spikes from PFR}
\item{\code{drops}}{Number of throws dropped from PFR}
\item{\code{drop_pct}}{Percentage of dropped throws from PFR}
\item{\code{bad_throws}}{Number of bad throws from PFR}
\item{\code{bad_throw_pct}}{Percentage of bad throws from PFR}
\item{\code{pocket_time}}{Average time in pocket from PFR}
\item{\code{times_blitzed}}{Number of times blitzed from PFR}
\item{\code{times_hurried}}{Number of times hurried from PFR}
\item{\code{times_hit}}{Number of times hit from PFR}
\item{\code{times_pressured}}{Number of times pressured from PFR}
\item{\code{pressure_pct}}{Percent of time pressured from PFR}
\item{\code{batted_balls}}{Number of batted balls from PFR}
\item{\code{on_tgt_throws}}{Number of throws on target from PFR}
\item{\code{on_tgt_pct}}{Percent of throws on target from PFR}
\item{\code{rpo_plays}}{Run-pass-option (RPO) number of plays from PFR}
\item{\code{rpo_yards}}{Run-pass-option (RPO) total yards from PFR}
\item{\code{rpo_pass_att}}{Run-pass-option (RPO) pass attempts from PFR}
\item{\code{rpo_pass_yards}}{Run-pass-option (RPO) pass yards from PFR}
\item{\code{rpo_rush_att}}{Run-pass-option (RPO) rush attempts from PFR}
\item{\code{rpo_rush_yards}}{Run-pass-option (RPO) rush yards from PFR}
\item{\code{pa_pass_att}}{Play action pass attempts from PFR}
\item{\code{pa_pass_yards}}{Play action pass yards from PFR}
\item{\code{passing_2pt_conversions}}{
Two-point conversion passes from PBP data
}
\item{\code{sacks}}{Total number of sacks taken from PBP data}
\item{\code{sack_yards}}{Total yards taken from sacks from PBP data}
\item{\code{sack_fumbles}}{Total fumbles from sacks from PBP data}
\item{\code{sack_fumbles_lost}}{Total fumbles lost from sacks from PBP data}
\item{\code{carries}}{
Number of rush attempts including scrambles and kneel downs. Rushes after
a lateral reception don't count as a carry from PBP data
}
\item{\code{rushing_yards}}{
Yards gained when rushing including scrambles and kneel downs. Also
includes yards gained after obtaining a lateral on a play that started
with a rushing attempt from PBP data
}
\item{\code{rushing_tds}}{
The number of rushing touchdowns (incl. scrambles). Also includes
touchdowns after obtaining a lateral on a play that started with a
rushing attempt from PBP data
}
\item{\code{rushing_fumbles}}{Number of rushes with a fumble from PBP data}
\item{\code{rushing_fumbles_lost}}{
Number of rushes with a lost fumble from PBP data
}
\item{\code{rushing_first_downs}}{
Number of rushing first downs from PBP data
}
\item{\code{rushing_epa}}{
Expected points added (EPA) on rush attempts including scrambles and
kneel downs from PBP data
}
\item{\code{rushing_2pt_conversions}}{
Two-point conversion rushes from PBP data
}
\item{\code{fpts_std_4pt_td}}{
Total fantasy points for standard scoring with 4 point touchdowns
}
\item{\code{ppg_std_4pt_td}}{
Points per game (PPG) for standard scoring with 4 point touchdowns
}
\item{\code{fpts_half_ppr_4pt_td}}{
Total fantasy points for half point per reception (Half PPR) scoring
with 4 point touchdowns
}
\item{\code{ppg_half_ppr_4pt_td}}{
Points per game (PPG) for half point per reception (Half PPR) scoring
with 4 point touchdowns
}
\item{\code{fpts_ppr_4pt_td}}{
Total fantasy points for point per reception (full PPR) scoring with
4 point touchdowns
}
\item{\code{ppg_ppr_4pt_td}}{
Points per game (PPG) for point per reception (Full PPR) scoring
with 4 point touchdowns
}
\item{\code{fpts_std_6pt_td}}{
Total fantasy points for standard scoring with 6 point touchdowns
}
\item{\code{ppg_std_6pt_td}}{
Points per game (PPG) for standard scoring with 6 point touchdowns
}
\item{\code{fpts_half_ppr_6pt_td}}{
Total fantasy points for half point per reception (Half PPR) scoring
with 6 point touchdowns
}
\item{\code{ppg_half_ppr_6pt_td}}{
Points per game (PPG) for half point per reception (Half PPR) scoring
with 6 point touchdowns
}
\item{\code{fpts_ppr_6pt_td}}{
Total fantasy points for point per reception (full PPR) scoring with
6 point touchdowns
}
\item{\code{ppg_ppr_6pt_td}}{
Points per game (PPG) for point per reception (Full PPR) scoring
with 6 point touchdowns
}
\item{\code{pfr_id}}{
Pro Football Reference player ID (e.g., JackLa00)
}
}
}
\usage{
get_qb_combined_stats_season(pbp_db = NULL, pbp_db_tbl = NULL, seasons = NULL)
}
\arguments{
\item{pbp_db_tbl}{Play-by-Play database table name (optional)}

\item{seasons}{NFL season (required) to obtain play-by-play data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}

\item{pbp_dp}{Play-by-Play database path (optional)}
}
\value{
Dataframe with QB stats for user-defined season(s) obtained from NFL
play-by-play data, Next Gen Stats (NGS) and Pro Football Reference (PFR)
}
\description{
Obtain QB stats from NFL cumulative season stats for a specified time frame
from either a saved database or \code{nflreadr::load_pbp()}. The stats are
obtained using play-by-play data, NFL Next Gen Stats (NGS) and Pro Football
Reference (PFR).
}
\details{
The function \code{get_qb_combined_stats_season} can be utilized to obtain
player stats for a season, including advanced stats.
This includes stats for passing, rushing, and receiving obtained by using
the \code{nflfastR} function \code{calculate_player_stats()}.
The player stats are utilized to calculate fantasy points
based on common scoring formats (4/6 point TD, STD, Half PPR, Full PPR).
The function acquires stats by using cumulative play-by-play data
by either loading from a saved database or using \code{nflreadr::load_pbp()}.
The data is obtained for a user-defined season.
Play-by-play data is merged with NFL Next Gen Stats (NGS) utilizing the
\code{nflreadr} function \code{load_nextgen_stats} to load player level weekly stats
starting with the 2016 season.
Play-by-play data is also merged with advanced stats from
Pro Football Reference (PFR), beginning from 2018.
Note that to use this function, \code{seasons} must be 2018 or later.
}
\seealso{
\code{\link[nuclearff]{nuclearff::get_pbp_data}}
Obtain play-by-play data for a specified time frame from either a saved
database or if not defined, using \code{nflreadr::load_pbp()}

\code{\link[nflreadr]{load_pbp}}
Load play-by-play data

\code{\link[nflfastR]{update_db}}
Update or Create a nflfastR Play-by-Play Database

\code{\link[nflreadr]{load_nextgen_stats}}
Load player level weekly NFL Next Gen Stats

\code{\link[nflreadr]{load_pfr_advstats}}
Load advanced stats from PFR
}
\author{
Nolan MacDonald
}
