% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pfr_advstats_season.R
\name{get_qb_pfr_advstats_season}
\alias{get_qb_pfr_advstats_season}
\title{QB Advanced Season Stats from Pro Football Reference (PFR)}
\format{
A data frame with 26 variables that are described below.
\describe{
\item{\code{player_display_name}}{Player name (e.g., Lamar Jackson)}
\item{\code{pass_attempts}}{Total pass attempts}
\item{\code{throwaways}}{Number of throwaways}
\item{\code{spikes}}{Number of spikes}
\item{\code{drops}}{Number of throws dropped}
\item{\code{drop_pct}}{Percentage of dropped throws}
\item{\code{bad_throws}}{Number of bad throws}
\item{\code{bad_throw_pct}}{Percentage of bad throws}
\item{\code{pocket_time}}{Average time in pocket}
\item{\code{times_blitzed}}{Number of times blitzed}
\item{\code{times_hurried}}{Number of times hurried}
\item{\code{times_hit}}{Number of times hit}
\item{\code{times_pressured}}{Number of times pressured}
\item{\code{pressure_pct}}{Percent of time pressured}
\item{\code{batted_balls}}{Number of batted balls}
\item{\code{on_tgt_throws}}{Number of throws on target}
\item{\code{on_tgt_pct}}{Percent of throws on target}
\item{\code{rpo_plays}}{Run-pass-option (RPO) number of plays}
\item{\code{rpo_yards}}{Run-pass-option (RPO) total yards}
\item{\code{rpo_pass_att}}{Run-pass-option (RPO) pass attempts}
\item{\code{rpo_pass_yards}}{Run-pass-option (RPO) pass yards}
\item{\code{rpo_rush_att}}{Run-pass-option (RPO) rush attempts}
\item{\code{rpo_rush_yards}}{Run-pass-option (RPO) rush yards}
\item{\code{pa_pass_att}}{Play action pass attempts}
\item{\code{pa_pass_yards}}{Play action pass yards}
\item{\code{pfr_player_id}}{
Pro Football Reference player ID (e.g., JackLa00)
}
}
}
\usage{
get_qb_pfr_advstats_season(seasons = NULL)
}
\arguments{
\item{seasons}{NFL Season(s) where \code{seasons = 2024}}
}
\value{
Dataframe with QB advanced stats from Pro Football Reference (PFR)
for a defined season
}
\description{
Obtain QB advanced season stats from \code{nflreadr},
to acquire Pro Football Reference (PFR) data.
}
\details{
\code{get_qb_pfr_advstats_season} is a function to obtain advanced stats from
Pro Football Reference (PFR).
The function utilizes the \code{nflreadr} function \code{load_pfr_advstats} to load
player level season stats starting with the 2018 season.
For quarterbacks, the \code{stat_type} is defined as \code{pass}.
More information is available about QB PFR advanced stats from
\href{https://nflreadr.nflverse.com/articles/dictionary_pfr_passing.html}{PFR passing data dictionary}.
}
\seealso{
\code{\link[nflreadr]{load_pfr_advstats}}
Load advanced stats from PFR
}
\author{
Nolan MacDonald
}
