% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_adv_stats.R
\name{get_rb_adv_stats_season}
\alias{get_rb_adv_stats_season}
\title{\strong{RB Cumulative Season Advanced Stats}}
\usage{
get_rb_adv_stats_season(
  pbp_db = "./data/pbp_db/pbp_db",
  pbp_db_tbl = "nflfastR_pbp",
  season = 2024
)
}
\arguments{
\item{pbp_db}{Path to play-by-play database and database name}

\item{pbp_db_tbl}{Play-by-play table name}

\item{season}{NFL Season}
}
\value{
Dataframe with RB advanced statistics for a defined season

\describe{
\item{\code{player}}{Player name}
\item{\code{team}}{Player team in defined season}
\item{\code{games}}{Total games played in defined season}
\item{\code{fantasy_points}}{Total fantasy points in a standard format (0 points per reception)}
\item{\code{fantasy_points_ppr}}{Total fantasy points in a full PPR format (1 point per reception)}
\item{\code{ppg}}{Average fantasy points per game}
}
}
\description{
Obtain RB cumulative season stats from \code{nflfastR}, and using \code{nflreadr}
to acquire Pro Football Reference (PFR) and Next Gen Stats (NGS) data
}
\details{
This is a function to obtain NFL advanced statistics for an entire
season. The function utilizes \code{nflfastR} for play-by-play data to determine
cumulative season statistics. The \code{nflreadr} package is also used to obtain
advanced statistics from Pro Football Reference (PFR) and
NFL Next Gen Stats (NGS). To use the function, you will need to save
play-by-play data to a database using the \code{nflfastR} function, \code{update_db()}.
For example, the database is saved by default as \code{pbp_db} with a table that
is stored containing all play-by-play information called \code{nflfastR_pbp}.
By defining the argument \code{pbp_db} (str) the user is specifying the path to the
database as well as the name it was saved as. To load the table inside the
database, \code{pbp_db_tbl} (str) can be defined if the naming is different,
however, the default argument follows the naming scheme provided by the
\code{nflfastR} package. The final requirement is the argument \code{season} (int)
to indicate the NFL season that the user wants to obtain advanced
statistics for.
More information is available on \code{update_db()} in the
\href{https://www.nflfastr.com/reference/update_db.html}{nflfastR documentation}.
}
\seealso{
\code{\link{get_snap_pct}}: Documentation on obtaining snap
share used in this function

\code{\link[nflfastR]{update_db}}: Create or update play-by-play
database

{\link[nflreadr]{load_pfr_advstats}}: PFR advanced statistics

{\link[nflreadr]{load_nextgen_stats}}: NFL Next Gen Stats
}
\author{
Nolan MacDonald
}
