% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_stats_season.R
\name{get_rb_combined_stats_season}
\alias{get_rb_combined_stats_season}
\title{RB Cumulative Season Advanced Stats}
\format{
A data frame with 57 variables that are described below.
\describe{
\item{\code{player_id}}{Player gsis id (e.g., 00-0034796)}
\item{\code{player_display_name}}{Player name (e.g., Lamar Jackson)}
\item{\code{player_name}}{Player shortened name (e.g., L.Jackson)}
\item{\code{pos}}{Player position (e.g., QB)}
\item{\code{tm}}{Player team (e.g., BAL)}
\item{\code{g}}{Number of games played}
\item{\code{carries}}{
Number of rush attempts including scrambles and kneel downs. Rushes after
a lateral reception don't count as a carry from PBP data
}
\item{\code{rushing_yards}}{
Yards gained when rushing including scrambles and kneel downs. Also
includes yards gained after obtaining a lateral on a play that started
with a rushing attempt from PBP data
}
\item{\code{rushing_tds}}{
The number of rushing touchdowns (incl. scrambles). Also includes
touchdowns after obtaining a lateral on a play that started with a
rushing attempt from PBP data
}
\item{\code{rushing_fumbles}}{Number of rushes with a fumble from PBP data}
\item{\code{rushing_fumbles_lost}}{
Number of rushes with a lost fumble from PBP data
}
\item{\code{rushing_first_downs}}{
Number of rushing first downs from PBP data
}
\item{\code{rushing_epa}}{
Expected points added (EPA) on rush attempts including scrambles and
kneel downs from PBP data
}
\item{\code{rushing_2pt_conversions}}{
Two-point conversion rushes from PBP data
}
\item{\code{efficiency}}{Average efficiency (EFF) from NGS}
\item{\code{percent_attempts_gte_eight_defenders }}{
Average percent attempts with 8+ Defenders in the Box (8+D) from NGS
}
\item{\code{avg_time_to_los}}{
Average time behind line of scrimmage (TLOS) from NGS
}
\item{\code{avg_rush_yards}}{Average rush yards per attempt}
\item{\code{expected_rush_yards}}{Average expected rush yards from NGS}
\item{\code{rush_yards_over_expected}}{
Average rush yards over expected from NGS
}
\item{\code{rush_yards_over_expected_per_att}}{
Average rush yards over expected per attempt from NGS
}
\item{\code{rush_pct_over_expected}}{
Average rush percent over expected from NGS
}
\item{\code{ybc}}{Rushing yards before contact from PFR}
\item{\code{ybc_att}}{
Rushing yards before contact per rushing attempt from PFR
}
\item{\code{yac}}{Rushing yards after contact from PFR}
\item{\code{yac_att}}{Rushing yards after contact per attempt from PFR}
\item{\code{brk_tkl}}{Broken tackles on rushes from PFR}
\item{\code{att_br}}{Rush attempts per broken tackle from PFR}
\item{\code{fpts_std_4pt_td}}{
Fantasy points for standard format with 4 point TD
}
\item{\code{ppg_std_4pt_td}}{
Fantasy points per game for standard format with 4 point TD
}
\item{\code{fpts_half_ppr_4pt_td}}{
Fantasy points for half PPR format with 4 point TD
}
\item{\code{ppg_half_ppr_4pt_td}}{
Fantasy points per game for half PPR format with 4 point TD
}
\item{\code{fpts_ppr_4pt_td}}{
Fantasy points for full PPR format with 4 point TD
}
\item{\code{ppg_ppr_4pt_td}}{
Fantasy points per game for full PPR format with 4 point TD
}
\item{\code{fpts_std_6pt_td}}{
Fantasy points for standard format with 6 point TD
}
\item{\code{ppg_std_6pt_td}}{
Fantasy points per game for standard format with 6 point TD
}
\item{\code{fpts_half_ppr_6pt_td}}{
Fantasy points for half PPR format with 6 point TD
}
\item{\code{ppg_half_ppr_6pt_td}}{
Fantasy points per game for half PPR format with 6 point TD
}
\item{\code{fpts_ppr_6pt_td}}{
Fantasy points for full PPR format with 6 point TD
}
\item{\code{ppg_ppr_6pt_td}}{
Fantasy points per game for full PPR format with 6 point TD
}
\item{\code{tgt}}{
Number of pass plays where the player was targeted as a receiver from
PBP data
}
\item{\code{rec}}{
Number of pass receptions. Lateral receptions don't count as a reception
from PBP data
}
\item{\code{rec_yds}}{
Yards gained after a pass reception. Includes yards gained after
receiving a lateral on a play that started as a pass play from PBP data
}
\item{\code{rec_td}}{
Number of reception touchdowns, including after receiving a lateral on a
play that began as a pass play from PBP data
}
\item{\code{receiving_fumbles}}{Number of fumbles after a pass reception
from PBP data
}
\item{\code{receiving_fumbles_lost}}{
Number of fumbles lost after a pass reception from PBP data
}
\item{\code{receiving_air}}{
Receiving air yards including incomplete passes from PBP data
}
\item{\code{yac_rec}}{
Yards after the catch gained on plays in which player was receiver (this
is an unofficial stat and may differ slightly between different sources)
from PBP data
}
\item{\code{x1d_rec}}{
Number of first downs gained on a reception from PBP data
}
\item{\code{receiving_epa}}{Expected points added on receptions
from PBP data
}
\item{\code{pfr_id}}{
Pro Football Reference player ID (e.g., JackLa00)
}
}
}
\usage{
get_rb_combined_stats_season(pbp_db = NULL, pbp_db_tbl = NULL, seasons = NULL)
}
\arguments{
\item{pbp_db_tbl}{Play-by-Play database table name (optional)}

\item{seasons}{NFL season (required) to obtain play-by-play data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}

\item{pbp_dp}{Play-by-Play database path (optional)}
}
\value{
Dataframe with QB stats for user-defined season(s) obtained from NFL
play-by-play data, Next Gen Stats (NGS) and Pro Football Reference (PFR)
}
\description{
Obtain RB stats from NFL cumulative season stats for a specified time frame
from either a saved database or \code{nflreadr::load_pbp()}. The stats are
obtained using play-by-play data, NFL Next Gen Stats (NGS) and Pro Football
Reference (PFR).
}
\details{
The function \code{get_rb_combined_stats_season} can be utilized to obtain
player stats for a season, including advanced stats.
This includes stats for passing, rushing, and receiving obtained by using
the \code{nflfastR} function \code{calculate_player_stats()}.
The player stats are utilized to calculate fantasy points
based on common scoring formats (4/6 point TD, STD, Half PPR, Full PPR).
The function acquires stats by using cumulative play-by-play data
by either loading from a saved database or using \code{nflreadr::load_pbp()}.
The data is obtained for a user-defined season.
Play-by-play data is merged with NFL Next Gen Stats (NGS) utilizing the
\code{nflreadr} function \code{load_nextgen_stats} to load player level weekly stats
starting with the 2016 season.
Play-by-play data is also merged with advanced stats from
Pro Football Reference (PFR), beginning from 2018.
Note that to use this function, \code{seasons} must be 2018 or later.
}
\seealso{
\code{\link[nuclearff]{nuclearff::get_pbp_data}}
Obtain play-by-play data for a specified time frame from either a saved
database or if not defined, using \code{nflreadr::load_pbp()}

\code{\link[nflreadr]{load_pbp}}
Load play-by-play data

\code{\link[nflfastR]{update_db}}
Update or Create a nflfastR Play-by-Play Database

\code{\link[nflreadr]{load_nextgen_stats}}
Load player level weekly NFL Next Gen Stats

\code{\link[nflreadr]{load_pfr_advstats}}
Load advanced stats from PFR
}
\author{
Nolan MacDonald
}
