% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pfr_advstats_season.R
\name{get_rb_pfr_advstats_season}
\alias{get_rb_pfr_advstats_season}
\title{RB Advanced Season Stats from Pro Football Reference (PFR)}
\format{
A data frame with 16 variables that are described below.
\describe{
\item{\code{player_display_name}}{Player name (e.g., Lamar Jackson)}
\item{\code{age}}{}
\item{\code{position}}{}
\item{\code{g}}{Number of games played}
\item{\code{gs}}{Number of games started}
\item{\code{att}}{Total rush attempts}
\item{\code{yds}}{Total rush yards}
\item{\code{td}}{Total rush touchdowns}
\item{\code{x1d}}{Rushing first downs}
\item{\code{ybc}}{Rushing yards before contact}
\item{\code{ybc_att}}{Rushing yards before contact per rushing attempt}
\item{\code{yac}}{Rushing yards after contact}
\item{\code{yac_att}}{Rushing yards after contact per attempt}
\item{\code{brk_tkl}}{Broken tackles on rushes}
\item{\code{att_br}}{Rush attempts per broken tackle}
\item{\code{pfr_player_id}}{
Pro Football Reference player ID (e.g., JackLa00)
}
}
}
\usage{
get_rb_pfr_advstats_season(seasons = NULL)
}
\arguments{
\item{seasons}{NFL Season(s) where \code{seasons = 2024}}
}
\value{
Dataframe with RB advanced stats from Pro Football Reference (PFR)
for a defined season
}
\description{
Obtain RB advanced season stats from \code{nflreadr},
to acquire Pro Football Reference (PFR) data.
}
\details{
\code{get_rb_pfr_advstats_season} is a function to obtain advanced stats from
Pro Football Reference (PFR).
The function utilizes the \code{nflreadr} function \code{load_pfr_advstats} to load
player level season stats starting with the 2018 season.
For running backs, the \code{stat_type} is defined as \code{rush}.
More information is available about RB PFR advanced stats from seasons on
the PFR site such as
\href{https://www.pro-football-reference.com/years/2024/rushing_advanced.htm}{PFR 2024 NFL Advanced Rushing}.
}
\seealso{
\code{\link[nflreadr]{load_pfr_advstats}}
Load advanced stats from PFR
}
\author{
Nolan MacDonald
}
