% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_gt.R
\name{get_scoring_format}
\alias{get_scoring_format}
\title{Helper Function - Get Scoring Format}
\usage{
get_scoring_format(scoring = NULL, position = NULL)
}
\arguments{
\item{scoring}{Fantasy scoring format (str). Can be defined as standard
scoring, half PPR scoring, full PPR scoring, 4PT PASS TD or 6PT PASS TD by
using: \code{"std_4pt_td"}, \code{"half_ppr_4pt_td"}, \code{"ppr_4pt_td"},
\code{"std_6pt_td"}, \code{"half_ppr_6pt_td"}, \code{"ppr_6pt_td"}}

\item{position}{Player position (str) to determine if extra format statement
is necessary (only in the case of QB for PASS TD points).}
}
\value{
String with fantasy scoring format to use in table subtitle
}
\description{
Obtain the valid columns to use in \code{dplyr::select} for a user-defined
scoring format.
}
\details{
Depending on the user-defined argument when creating \code{gt} tables for
fantasy stats, the scoring format columns can be included while neglecting
all other calculated fantasy scoring formats.
}
\seealso{
\code{\link[nuclearff]{calc_fpts_common_formats}}
Calculate fantasy points based on common scoring formats

\code{\link[nuclearff]{calc_fpts_ppg_common_formats}}
Calculate fantasy points per game (PPG) based on common scoring formats

\code{\link[nuclearff]{table_ovr_qb_fantasy}}
Create table with overall fantasy standings for QB
}
\author{
Nolan MacDonald
}
