% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_pct.R
\name{get_snap_pct}
\alias{get_snap_pct}
\title{Player Snap Percentages}
\usage{
get_snap_pct(szn = 2024, pos = "WR")
}
\arguments{
\item{szn}{Season}

\item{pos}{Player position}
}
\value{
Obtain player snap share (\%) for positional group
}
\description{
Player Snap Percentages
}
