% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snap_share.R
\name{get_snap_share}
\alias{get_snap_share}
\title{Player Snap Share}
\usage{
get_snap_share(season = NULL, pos = NULL)
}
\arguments{
\item{pos}{Player position (str)}

\item{seasons}{NFL Season(s) (int) where \code{seasons = 2024} or
\code{seasons = c(2023, 2024)}}
}
\value{
Obtain player snap share percentage for positional group
}
\description{
Obtain player snap shares provided by Pro Football Reference (PFR),
starting with the 2012 season.
}
\details{
This function loads game level snap counts by PFR starting with the 2012
season. Snap count data is then utilized to filter by season and player
position. Additional data wrangling is performed to format player names
in order to merge easily with play-by-play data.
More information is available about snap counts from
\href{https://nflreadr.nflverse.com/reference/load_snap_counts.html}{Load Snap Counts from PFR}.
}
\seealso{
\code{\link[nflreadr]{load_players}}
Obtain player information from \code{nflreadr}

\code{\link[nflreadr]{load_snap_counts}}
Load game level snap counts stats provided by PFR
}
\author{
Nolan MacDonald

\itemize{
\item{\code{player}}{Player name (e.g., Darnell Mooney)}
\item{\code{position}}{Player name (e.g., WR)}
\item{\code{team}}{Player team}
\item{\code{pfr_player_id}}{Pro Football Reference (PFR) player ID}
\item{\code{snaps}}{Number of snaps}
\item{\code{snap_share}}{Fraction of total snaps}
\item{\code{snap_pct}}{Percentage of total snaps}
}
}
