% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_stats_season.R
\name{get_wr_combined_stats_season}
\alias{get_wr_combined_stats_season}
\title{WR Cumulative Season Advanced Stats}
\format{
A data frame with 61 variables that are described below.
\describe{
\item{\code{player_id}}{Player gsis id (e.g., 00-0034796)}
\item{\code{player_display_name}}{Player name (e.g., Lamar Jackson)}
\item{\code{player_name}}{Player shortened name (e.g., L.Jackson)}
\item{\code{pos}}{Player position (e.g., QB)}
\item{\code{tm}}{Player team (e.g., BAL)}
\item{\code{g}}{Number of games played}
\item{\code{tgt}}{
Number of pass plays where the player was targeted as a receiver from
PBP data
}
\item{\code{rec}}{
Number of pass receptions. Lateral receptions don't count as a reception
from PBP data
}
\item{\code{rec_yds}}{
Yards gained after a pass reception. Includes yards gained after
receiving a lateral on a play that started as a pass play from PBP data
}
\item{\code{rec_td}}{
Number of reception touchdowns, including after receiving a lateral on a
play that began as a pass play from PBP data
}
\item{\code{avg_yac}}{Average yards after catch (YAC) from NGS}
\item{\code{avg_expected_yac}}{Average expected yards after catch (xYAC) from NGS}
\item{\code{avg_yac_above_expectation}}{
Average yards after catch above expectation (+/-) from NGS
}
\item{\code{yac}}{Yards after catch (YAC) from PFR}
\item{\code{yac_r}}{Yards after catch (YAC) per reception from PFR}
\item{\code{adot}}{
Average depth of target (ADOT) when targeted, whether completed or not,
from PFR
}
\item{\code{receiving_epa}}{Expected points added on receptions
from PBP data
}
\item{\code{racr}}{
Receiving Air Conversion Ratio.
RACR = \code{receiving_yards} / \code{receiving_air_yards} from PBP
}
\item{\code{wopr}}{
Weighted Opportunity Rating.
WOPR = 1.5 x \code{target_share} + 0.7 x \code{air_yards_share} from PBP data
}
\item{\code{avg_cushion}}{Average cushion (CUSH) from NGS}
\item{\code{avg_separation}}{Average separation (SEP) from NGS}
\item{\code{target_share}}{
Share of targets of player compared to all team targets from PBP
}
\item{\code{tgt_pct}}{Share of targets percentage from PBP}
\item{\code{ybc}}{
Total yards passes traveled in the air before being caught or yards
before catch (YBC) from PFR}
\item{\code{ybc_r}}{Yards before catch per reception from PFR}
\item{\code{receiving_air}}{
Receiving air yards including incomplete passes from PBP data
}
\item{\code{avg_intended_air_yards}}{Average targeted air yards (TAY) from NGS}
\item{\code{percent_share_of_intended_air_yards}}{
Average \% share of team's air yards (TAY\%) from NGS
}
\item{\code{air_yards_share}}{
Share of \code{receiving_air_yards} of the player to all team \code{air_yards}
from PBP data
}
\item{\code{x1d_rec}}{
Number of first downs gained on a reception from PBP data
}
\item{\code{brk_tkl}}{Number of broken tackles from PFR}
\item{\code{rec_br}}{Receptions per broken tackle from PFR}
\item{\code{drop}}{Number of dropped passes from PFR}
\item{\code{drop_percent}}{Dropped pass percentage when targeted from PFR}
\item{\code{catch_percentage}}{Average catch percentage from NGS}
\item{\code{receiving_fumbles}}{Number of fumbles after a pass reception
from PBP data
}
\item{\code{receiving_fumbles_lost}}{
Number of fumbles lost after a pass reception from PBP data
}
\item{\code{receiving_2pt_conversions}}{
Two-point conversion receptions from PBP data
}
\item{\code{int_tgt}}{Interceptions on passes where targetedfrom PFR}
\item{\code{rat}}{Passer rating on passes when targeted from PFR}
\item{\code{fpts_std_4pt_td}}{
Fantasy points for standard format with 4 point TD
}
\item{\code{ppg_std_4pt_td}}{
Fantasy points per game for standard format with 4 point TD
}
\item{\code{fpts_half_ppr_4pt_td}}{
Fantasy points for half PPR format with 4 point TD
}
\item{\code{ppg_half_ppr_4pt_td}}{
Fantasy points per game for half PPR format with 4 point TD
}
\item{\code{fpts_ppr_4pt_td}}{
Fantasy points for full PPR format with 4 point TD
}
\item{\code{ppg_ppr_4pt_td}}{
Fantasy points per game for full PPR format with 4 point TD
}
\item{\code{fpts_std_6pt_td}}{
Fantasy points for standard format with 6 point TD
}
\item{\code{ppg_std_6pt_td}}{
Fantasy points per game for standard format with 6 point TD
}
\item{\code{fpts_half_ppr_6pt_td}}{
Fantasy points for half PPR format with 6 point TD
}
\item{\code{ppg_half_ppr_6pt_td}}{
Fantasy points per game for half PPR format with 6 point TD
}
\item{\code{fpts_ppr_6pt_td}}{
Fantasy points for full PPR format with 6 point TD
}
\item{\code{ppg_ppr_6pt_td}}{
Fantasy points per game for full PPR format with 6 point TD
}
\item{\code{carries}}{
Number of rush attempts including scrambles and kneel downs. Rushes after
a lateral reception don't count as a carry from PBP data
}
\item{\code{rushing_yards}}{
Yards gained when rushing including scrambles and kneel downs. Also
includes yards gained after obtaining a lateral on a play that started
with a rushing attempt from PBP data
}
\item{\code{rushing_tds}}{
The number of rushing touchdowns (incl. scrambles). Also includes
touchdowns after obtaining a lateral on a play that started with a
rushing attempt from PBP data
}
\item{\code{rushing_fumbles}}{Number of rushes with a fumble from PBP data}
\item{\code{rushing_fumbles_lost}}{
Number of rushes with a lost fumble from PBP data
}
\item{\code{rushing_first_downs}}{
Number of rushing first downs from PBP data
}
\item{\code{rushing_epa}}{
Expected points added (EPA) on rush attempts including scrambles and
kneel downs from PBP data
}
\item{\code{rushing_2pt_conversions}}{
Two-point conversion rushes from PBP data
}
\item{\code{pfr_id}}{
Pro Football Reference player ID (e.g., JackLa00)
}
}
}
\usage{
get_wr_combined_stats_season(pbp_db = NULL, pbp_db_tbl = NULL, seasons = NULL)
}
\arguments{
\item{pbp_db_tbl}{Play-by-Play database table name (optional)}

\item{seasons}{NFL season (required) to obtain play-by-play data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}

\item{pbp_dp}{Play-by-Play database path (optional)}
}
\value{
Dataframe with QB stats for user-defined season(s) obtained from NFL
play-by-play data, Next Gen Stats (NGS) and Pro Football Reference (PFR)
}
\description{
Obtain WR stats from NFL cumulative season stats for a specified time frame
from either a saved database or \code{nflreadr::load_pbp()}. The stats are
obtained using play-by-play data, NFL Next Gen Stats (NGS) and Pro Football
Reference (PFR).
}
\details{
The function \code{get_wr_combined_stats_season} can be utilized to obtain
player stats for a season, including advanced stats.
This includes stats for passing, rushing, and receiving obtained by using
the \code{nflfastR} function \code{calculate_player_stats()}.
The player stats are utilized to calculate fantasy points
based on common scoring formats (4/6 point TD, STD, Half PPR, Full PPR).
The function acquires stats by using cumulative play-by-play data
by either loading from a saved database or using \code{nflreadr::load_pbp()}.
The data is obtained for a user-defined season.
Play-by-play data is merged with NFL Next Gen Stats (NGS) utilizing the
\code{nflreadr} function \code{load_nextgen_stats} to load player level weekly stats
starting with the 2016 season.
Play-by-play data is also merged with advanced stats from
Pro Football Reference (PFR), beginning from 2018.
Note that to use this function, \code{seasons} must be 2018 or later.
}
\seealso{
\code{\link[nuclearff]{nuclearff::get_pbp_data}}
Obtain play-by-play data for a specified time frame from either a saved
database or if not defined, using \code{nflreadr::load_pbp()}

\code{\link[nflreadr]{load_pbp}}
Load play-by-play data

\code{\link[nflfastR]{update_db}}
Update or Create a nflfastR Play-by-Play Database

\code{\link[nflreadr]{load_nextgen_stats}}
Load player level weekly NFL Next Gen Stats

\code{\link[nflreadr]{load_pfr_advstats}}
Load advanced stats from PFR
}
\author{
Nolan MacDonald
}
