% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ngs_advstats_season.R
\name{get_wr_ngs_advstats_season}
\alias{get_wr_ngs_advstats_season}
\title{WR Cumulative Season NFL Next Gen Stats (NGS)}
\format{
A data frame with 16 variables that are described below.
\describe{
\item{\code{player_id}}{Player gsis id (e.g., 00-0039337)}
\item{\code{player_display_name}}{Player name (e.g., Malik Nabers)}
\item{\code{player_name}}{Player shortened name (e.g., M.Nabers)}
\item{\code{position}}{Player position}
\item{\code{targets}}{Total receiving targets}
\item{\code{receptions}}{Total receptions}
\item{\code{yards}}{Total receiving yards}
\item{\code{rec_touchdowns}}{Total reception touchdowns}
\item{\code{avg_cushion}}{Average cushion (CUSH)}
\item{\code{avg_separation}}{Average separation (SEP)}
\item{\code{avg_intended_air_yards}}{Average targeted air yards (TAY)}
\item{\code{percent_share_of_intended_air_yards}}{
Average \% share of team's air yards (TAY\%)
}
\item{\code{catch_percentage}}{Average catch percentage}
\item{\code{avg_yac}}{Average yards after catch (YAC)}
\item{\code{avg_expected_yac}}{Average expected yards after catch (xYAC)}
\item{\code{avg_yac_above_expectation}}{
Average yards after catch above expectation (+/-)
}
}
}
\usage{
get_wr_ngs_advstats_season(seasons = NULL)
}
\arguments{
\item{seasons}{NFL Season(s) where \code{seasons = 2024} or
\code{seasons = c(2023, 2024)}}
}
\value{
Dataframe with WR Next Gen Stats for a defined season
}
\description{
Obtain WR cumulative season stats from \code{nflreadr},
to acquire Next Gen Stats (NGS) data.
}
\details{
\code{get_wr_ngs_advstats_season} is a function to obtain NFL Next Gen
Stats (NGS) for an entire season or multiple seasons.
The function utilizes the \code{nflreadr} function \code{load_nextgen_stats} to load
player level weekly stats starting with the 2016 season.
For receivers, the \code{stat_type} is defined as \code{receiving} and is updated
every night.
NGS will only provide data for players above a minimum number of receptions.
More information is available about NGS stats from
\href{https://nextgenstats.nfl.com/glossary}{NGS glossary}.
}
\seealso{
\code{\link[nflreadr]{load_nextgen_stats}}
Load player level weekly NFL Next Gen Stats
}
\author{
Nolan MacDonald
}
