% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_data_from_db}
\alias{load_data_from_db}
\title{Helper - Load Play-by-Play Data from Database}
\usage{
load_data_from_db(pbp_db, pbp_db_tbl, season, week_min, week_max)
}
\arguments{
\item{pbp_db_tbl}{Play-by-Play database table name}

\item{season}{NFL season to obtain play-by-play data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}

\item{week_min}{Minimum week to define whether pulling a range
of weeks or the entire season. Use \code{week_min = 1} for the first week of
the season, must be an integer.}

\item{week_max}{Maximum week to define a range of weeks to pull
from an NFL season. If not defined, the data will be pulled for all weeks,
beginning with \code{week_min}.}

\item{pbp_dp}{Play-by-Play database path}
}
\value{
Dataframe containing filtered play-by-play data
}
\description{
Helper function to load play-by-play data from \code{nflreadr} for a specified
time frame.
}
\details{
Helper function to load play-by-play data from from a database. The function
filters the data to include a specified time frame including seasons
and a range of weeks.
If \code{week_max} is provided, filter between \code{week_min} and \code{week_max}
If \code{week_max} is not provided, filter for \code{weeks >= week_min} to get the
remainder of the season.
This function is typically utilized if the user
desires loading play-by-play data from a database and defined \code{pbp_db} and
\code{pbp_db_tbl} as strings.
}
\seealso{
\code{\link[nflfastR]{update_db}}:
Update or Create a nflfastR Play-by-Play Database

\code{\link[nuclearff]{get_pbp_data}}:
Obtain filtered play-by-play data utilizes this helper function
}
\author{
Nolan MacDonald
}
