% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_stats.R
\name{load_data_from_nflreadr}
\alias{load_data_from_nflreadr}
\title{Helper - Load Play-by-Play Data from \code{nflreadr}}
\usage{
load_data_from_nflreadr(season, week_min, week_max)
}
\arguments{
\item{season}{NFL season to obtain play-by-play data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}

\item{week_min}{Minimum week to define whether pulling a range
of weeks or the entire season. Use \code{week_min = 1} for the first week of
the season, must be an integer.}

\item{week_max}{Maximum week to define a range of weeks to pull
from an NFL season. If not defined, the data will be pulled for all weeks,
beginning with \code{week_min}.}
}
\value{
Dataframe containing filtered play-by-play data
}
\description{
Helper function to load play-by-play data from \code{nflreadr} for a specified
time frame.
}
\details{
Helper function to load play-by-play data from \code{nflreadr}. The function
filters the data to include a specified time frame including seasons
and a range of weeks.
If \code{week_max} is provided, filter between \code{week_min} and \code{week_max}
If \code{week_max} is not provided, filter for \code{weeks >= week_min} to get the
remainder of the season.
This function is typically utilized if the user is
not loading play-by-play data from a database.
}
\seealso{
\code{\link[nflreadr]{load_pbp}}:
Load play-by-play data

\code{\link[nuclearff]{get_pbp_data}}:
Obtain filtered play-by-play data utilizes this helper function
}
\author{
Nolan MacDonald
}
