% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rosters.R
\name{replace_player_names}
\alias{replace_player_names}
\title{Helper - Cleanup Player Names}
\usage{
replace_player_names(df)
}
\arguments{
\item{df}{Dataframe with player column to clean up names}
}
\value{
Dataframe with cleaned up player names
}
\description{
String parsing helper to cleanup player names for continuity.
}
\details{
The function \code{replace_player_names} reads a dataframe that includes player
names and uses \code{str_replace} to clean up names. This can be used across the
package when joining data such that all player names are the same.
The function strips names of suffixes like Jr., Sr., II, III. Also, the
first names that are abbreviations like D.J. Moore and D.K. Metcalf have
the periods stripped from the initials. All other player names are
specified to say the same. Some specific cases are defined explicitly.
}
\author{
Nolan MacDonald
}
