% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rosters.R
\name{replace_team_names}
\alias{replace_team_names}
\title{Helper - Cleanup Team Abbreviations}
\usage{
replace_team_names(df)
}
\arguments{
\item{df}{Dataframe with team column to clean up abbreviations}
}
\value{
Dataframe with cleaned up team abbreviations
}
\description{
String parsing helper to cleanup team abbreviations for continuity.
}
\details{
The function \code{team_player_names} reads a dataframe that includes team
names as \code{team} and uses \code{str_replace} to replace the team abbreviation.
This can be used across the package when joining data such that all team
names are the same. Team specific cases are defined explicitly, to
alleviate the issue of some data differences. For example, \code{nflreadr} roster
data lists the Los Angeles Rams as LA, while \code{nflfastR} play-by-play data
lists the abbreviation as LAR.
}
\author{
Nolan MacDonald
}
