% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_gt.R
\name{save_gt_table}
\alias{save_gt_table}
\title{Helper Function - Save Table}
\usage{
save_gt_table(table = NULL, file_name = NULL, output_dir = NULL)
}
\arguments{
\item{table}{Name of the variable for the table to save}

\item{file_name}{Name of the file to save the table as, typically a \code{.png} file (str)}

\item{output_dir}{Directory to save the table to (str)}
}
\value{
description
}
\description{
Save a \code{gt} table to a file in the specified directory.
}
\details{
Save a \code{gt} table to a file in the specified directory. The file name and output directory must be provided.
If the output directory does not exist, it will be created.
}
\seealso{
\code{\link[nuclearff]{nuclearff::get_wr_pbp_stats}}
WR cumulative season stats obtained from play-by-play data

\code{\link[nuclearff]{nuclearff::get_player_data}}
Get player data and IDs to merge dataframes

\code{\link[nuclearff]{nuclearff::get_snap_share}}
Pull player snap shares
}
\author{
Nolan MacDonald
}
