% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gt_ovr_fantasy_standings.R
\name{table_ovr_rb_fantasy}
\alias{table_ovr_rb_fantasy}
\title{RB Overall Fantasy Football Standings Table}
\usage{
table_ovr_rb_fantasy(
  num_players = NULL,
  scoring = NULL,
  seasons = NULL,
  pbp_db = NULL,
  pbp_db_tbl = NULL,
  file_name = NULL,
  output_dir = NULL
)
}
\arguments{
\item{num_players}{Number of players to list in table (int)}

\item{scoring}{Fantasy scoring format (str). Can be defined as standard
scoring, half PPR scoring, full PPR scoring, 4PT PASS TD or 6PT PASS TD by
using: \code{"std_4pt_td"}, \code{"half_ppr_4pt_td"}, \code{"ppr_4pt_td"},
\code{"std_6pt_td"}, \code{"half_ppr_6pt_td"}, \code{"ppr_6pt_td"}}

\item{seasons}{NFL season or year (int) such as \code{seasons = 2024}.}

\item{pbp_db}{Path to play-by-play database (str)}

\item{pbp_db_tbl}{Table name stored in the play-by-play database (str)}

\item{file_name}{File name to save table}

\item{output_dir}{Directory to save the table}
}
\value{
Rendered table using \code{gt} and a saved \code{.png} if \code{output_dir} and
\code{file_name} are defined
}
\description{
Create a table using \code{gt} of overall player standings in fantasy football.
The table is colored by fantasy points per game (PPG) and ordered by total
fantasy points.
}
\details{
Using \code{nflfastR} play-by-play data, stats for a user-defined season can
be obtained to create a table with overall standings for a season.
The function automatically includes the entire season, defined by year as
an integer. The table is colored by fantasy points per game (PPG) using the
\code{gtExtras} hulk theme and ordered by total fantasy points.
The table is saved as a \code{.png} figure.
}
\seealso{
\code{\link[nuclearff]{nuclearff::get_rb_pbp_stats}}
RB cumulative season stats obtained from play-by-play data

\code{\link[nuclearff]{nuclearff::get_player_data}}
Get player data and IDs to merge dataframes

\code{\link[nuclearff]{nuclearff::get_snap_share}}
Pull player snap shares
}
\author{
Nolan MacDonald
}
