% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_advstats.R
\name{validate_ngs_season}
\alias{validate_ngs_season}
\title{Helper - Validate NFL Next Gen Stats (NGS) Data Seasons}
\usage{
validate_ngs_season(season)
}
\arguments{
\item{season}{NFL season to obtain NGS data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}
}
\value{
Validate if \code{season} is defined as an integer of 2016 or greater
}
\description{
Validation helper to determine if the seasons argument to obtain NGS is
2016 or later.
}
\details{
The function \code{validate_ngs_season} can be utilized to validate whether the
arguments for \code{season} are appropriate as NFL NGS dates back
to 2016. If any season is defined prior to 2016, an error will occur.
}
\seealso{
\code{\link[nuclearff]{get_qb_ngs_advstats_season}}:
Obtain QB cumulative season stats from \code{nflreadr}, to acquire
Next Gen Stats (NGS) data. The function utilizes this helper.

\code{\link[nuclearff]{get_rb_ngs_advstats_season}}:
Obtain RB cumulative season stats from \code{nflreadr}, to acquire
Next Gen Stats (NGS) data. The function utilizes this helper.

\code{\link[nuclearff]{get_wr_ngs_advstats_season}}:
Obtain WR cumulative season stats from \code{nflreadr}, to acquire
Next Gen Stats (NGS) data. The function utilizes this helper.

\code{\link[nuclearff]{get_te_ngs_advstats_season}}:
Obtain TE cumulative season stats from \code{nflreadr}, to acquire
Next Gen Stats (NGS) data. The function utilizes this helper.
}
\author{
Nolan MacDonald
}
