% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_stats.R
\name{validate_pbp_db}
\alias{validate_pbp_db}
\title{Helper - Validate Play-by-Play Database Path and Table Arguments}
\usage{
validate_pbp_db(pbp_db, pbp_db_tbl)
}
\arguments{
\item{pbp_db_tbl}{Play-by-Play database table name (str)}

\item{pbp_dp}{Play-by-Play database path (str)}
}
\value{
Validate if \code{pbp_db} and \code{pbp_db_tbl} are strings
}
\description{
Validation helper to determine if the path to the play-by-play database,
\code{pbp_db}, and the database table name, \code{pbp_db_tbl}, are defined as strings.
}
\details{
The function \code{validate_pbp_db} can be utilized to validate whether the
arguments to load a database in \code{get_pbp_data} are valid and defined as
strings.
}
\seealso{
\code{\link[nuclearff]{get_pbp_data}}:
Obtain filtered play-by-play data utilizes this helper function
}
\author{
Nolan MacDonald
}
