% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_stats.R
\name{validate_pbp_season}
\alias{validate_pbp_season}
\title{Helper - Validate Play-by-Play Data Seasons}
\usage{
validate_pbp_season(season)
}
\arguments{
\item{season}{NFL season to obtain play-by-play data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}
}
\value{
Validate if \code{season} is defined as an integer of 1999 or greater
}
\description{
Validation helper to determine if the seasons argument to obtain the
play-by-play database is 1999 or later.
}
\details{
The function \code{validate_pbp_season} can be utilized to validate whether the
arguments for \code{season} are appropriate as NFL play-by-play data dates back
to 1999. If any season is defined prior to 1999, an error will occur.
}
\seealso{
\code{\link[nuclearff]{get_pbp_data}}:
Obtain filtered play-by-play data utilizes this helper function
}
\author{
Nolan MacDonald
}
