% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_stats.R
\name{validate_pbp_weeks}
\alias{validate_pbp_weeks}
\title{Helper - Validate Arguments for Weeks}
\usage{
validate_pbp_weeks(week_min, week_max)
}
\arguments{
\item{week_min}{Minimum week to define whether pulling a range
of weeks or the entire season. Use \code{week_min = 1} for the first week of
the season, must be an integer.}

\item{week_max}{Maximum week to define a range of weeks to pull
from an NFL season. If not defined, the data will be pulled for all weeks,
beginning with \code{week_min}.}
}
\value{
Validate if weeks are defined as integers
}
\description{
Validation helper to determine if weeks are defined properly to filter the
play-by-play data.
}
\details{
The function \code{validate_pbp_weeks} is utilized to determine weeks are
defined as integers for \code{week_min} and \code{week_max}.
}
\seealso{
\code{\link[nuclearff]{get_pbp_data}}:
Obtain filtered play-by-play data utilizes this helper function
}
\author{
Nolan MacDonald
}
