% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_advstats.R
\name{validate_pfr_season}
\alias{validate_pfr_season}
\title{Helper - Validate Pro Football Reference (PFR) Advanced Stats Seasons}
\usage{
validate_pfr_season(season)
}
\arguments{
\item{season}{NFL season to obtain NGS data. The
season can be defined as a single season, \code{season = 2024}. For multiple
seasons, use either \code{season = c(2023,2024)} or \code{season = 2022:2024}.}
}
\value{
Validate if \code{season} is defined as an integer of 2018 or greater
}
\description{
Validation helper to determine if the seasons argument to obtain PFR is
2018 or later.
}
\details{
The function \code{validate_pfr_season} can be utilized to validate whether the
arguments for \code{season} are appropriate as PFR advanced stats dates back
to 2018. If any season is defined prior to 2018, an error will occur.
}
\seealso{
\code{\link[nuclearff]{get_qb_pfr_advstats_season}}:
Obtain QB advanced season stats from \code{nflreadr}, to acquire Pro Football
Reference (PFR) data. The function utilizes this helper.

\code{\link[nuclearff]{get_rb_pfr_advstats_season}}:
Obtain RB advanced season stats from \code{nflreadr}, to acquire Pro Football
Reference (PFR) data. The function utilizes this helper.

\code{\link[nuclearff]{get_wr_pfr_advstats_season}}:
Obtain WR advanced season stats from \code{nflreadr}, to acquire Pro Football
Reference (PFR) data. The function utilizes this helper.

\code{\link[nuclearff]{get_te_pfr_advstats_season}}:
Obtain TE advanced season stats from \code{nflreadr}, to acquire Pro Football
Reference (PFR) data. The function utilizes this helper.
}
\author{
Nolan MacDonald
}
