% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fpts.R
\name{calc_fpts_common_formats}
\alias{calc_fpts_common_formats}
\title{Calculate Fantasy Points for Common Scoring Formats}
\usage{
calc_fpts_common_formats(pbp = pbp)
}
\arguments{
\item{pbp}{Play-by-play Dataframe}
}
\value{
Calculate fantasy points using play-by-play and data and common
scoring settings (i.e., 4/6PT PASS TD, -2 PASS INT, STD/Half PPR/Full PPR)
}
\description{
Calculate Fantasy Points for Common Scoring Formats
}
